extract command step to commons
