+++ packages/allure-cucumberjs/src/CucumberAllureWorld.ts
@@ -1,16 +1,12 @@
 import { World } from "@cucumber/cucumber";
 import {
-  Attachment,
+  AllureCommandStep,
+  AllureCommandStepExecutable,
   AttachmentMetadata,
-  AttachmentOptions,
   Category,
-  ContentType,
-  ExecutableItem,
   LabelName,
   LinkType,
   ParameterOptions,
-  Stage,
-  Status,
 } from "allure-js-commons";
 import { ALLURE_METADATA_CONTENT_TYPE } from "allure-js-commons/internal";
 
@@ -19,211 +15,10 @@ export interface CucumberAttachmentMetadata extends AttachmentMetadata {
   categories?: Category[];
 }
 
-export interface AllureCommandStep<T = AttachmentMetadata> {
-  name: string;
-
-  attachments: Attachment[];
-
-  metadata: T;
-
-  label(label: string, value: string): void | Promise<void>;
-
-  link(url: string, name?: string, type?: string): void | Promise<void>;
-
-  parameter(name: string, value: string, options?: ParameterOptions): void | Promise<void>;
-
-  epic(epic: string): void | Promise<void>;
-
-  feature(feature: string): void | Promise<void>;
-
-  story(story: string): void | Promise<void>;
-
-  suite(name: string): void | Promise<void>;
-
-  parentSuite(name: string): void | Promise<void>;
-
-  subSuite(name: string): void | Promise<void>;
-
-  owner(owner: string): void | Promise<void>;
-
-  severity(severity: string): void | Promise<void>;
-
-  tag(tag: string): void | Promise<void>;
-
-  issue(issue: string, url: string): void | Promise<void>;
-
-  tms(issue: string, url: string): void | Promise<void>;
-
-  attach(
-    name: string,
-    content: Buffer | string,
-    options: ContentType | string | AttachmentOptions,
-  ): void | Promise<void>;
-}
-
-// TODO: think about the class name
-export class AllureCommandStepExecutable implements AllureCommandStep {
-  name: string = "";
-
-  attachments: Attachment[] = [];
-
-  metadata: CucumberAttachmentMetadata = {};
-
-  constructor(name: string) {
-    this.name = name;
-  }
-
-  label(label: string, value: string): void {
-    if (!this.metadata.labels) {
-      this.metadata.labels = [];
-    }
-
-    this.metadata.labels.push({
-      name: label,
-      value,
-    });
-  }
-
-  link(url: string, name?: string, type?: string): void {
-    if (!this.metadata.links) {
-      this.metadata.links = [];
-    }
-
-    this.metadata.links.push({
-      name,
-      url,
-      type,
-    });
-  }
-
-  parameter(name: string, value: string, options?: ParameterOptions): void {
-    if (!this.metadata.parameter) {
-      this.metadata.parameter = [];
-    }
-
-    this.metadata.parameter.push({
-      name,
-      value,
-      hidden: options?.hidden || false,
-      excluded: options?.excluded || false,
-    });
-  }
-
-  epic(epic: string) {
-    this.label(LabelName.EPIC, epic);
-  }
-
-  feature(feature: string) {
-    this.label(LabelName.FEATURE, feature);
-  }
-
-  story(story: string) {
-    this.label(LabelName.STORY, story);
-  }
-
-  suite(name: string) {
-    this.label(LabelName.SUITE, name);
-  }
-
-  parentSuite(name: string) {
-    this.label(LabelName.PARENT_SUITE, name);
-  }
-
-  subSuite(name: string) {
-    this.label(LabelName.SUB_SUITE, name);
-  }
-
-  owner(owner: string) {
-    this.label(LabelName.OWNER, owner);
-  }
-
-  severity(severity: string) {
-    this.label(LabelName.SEVERITY, severity);
-  }
-
-  tag(tag: string) {
-    this.label(LabelName.TAG, tag);
-  }
-
-  issue(name: string, url: string) {
-    this.link(url, name, LinkType.ISSUE);
-  }
-
-  tms(name: string, url: string) {
-    this.link(url, name, LinkType.TMS);
-  }
-
-  attach(source: string | Buffer, type: string): void {
-    console.log("attach", source);
-
-    this.attachments.push({
-      name: "attachment",
-      source: Buffer.isBuffer(source)
-        ? source.toString("base64")
-        : Buffer.from(source, "utf8").toString("base64"),
-      type,
-    });
-  }
-
-  async start(
-    body: (step: AllureCommandStepExecutable) => any | Promise<any>,
-  ): Promise<AttachmentMetadata> {
-    const startDate = new Date().getTime();
-
-    try {
-      const res = body.call(this, this);
-      const stepResult = await res;
-
-      console.log("step metadata", this.metadata);
-
-      return {
-        ...this.metadata,
-        steps: [
-          {
-            name: this.name,
-            start: startDate,
-            stop: new Date().getTime(),
-            stage: Stage.FINISHED,
-            status: Status.PASSED,
-            statusDetails: {},
-            attachments: this.attachments,
-            parameters: [],
-            steps: [],
-          },
-        ],
-      };
-    } catch (err) {
-      return {
-        ...this.metadata,
-        steps: [
-          {
-            name: this.name,
-            start: startDate,
-            stop: new Date().getTime(),
-            stage: Stage.FINISHED,
-            status: Status.FAILED,
-            statusDetails:
-              err instanceof Error
-                ? {
-                    message: err.message,
-                    trace: err.stack,
-                  }
-                : {},
-            attachments: this.attachments,
-            parameters: [],
-            steps: [],
-          },
-        ],
-      };
-    }
-  }
-}
-
 // TODO: check implemetation details
 export class CucumberAllureWorld
   extends World
-  implements Omit<AllureCommandStep, "attach" | "name" | "attachments" | "metadata">
-{
+  implements Omit<AllureCommandStep, "attach" | "name" | "attachments" | "metadata"> {
   public async label(label: string, value: string) {
     const msgBody: AttachmentMetadata = {
       labels: [
